Refactoring TextBlock and its minions
* Shoes3/shoes3#333
