+++ shoes/types/types.h
@@ -18,6 +18,7 @@
 #include "shoes/types/text.h"
 #include "shoes/types/text_link.h"
 #include "shoes/types/text_view.h"
+#include "shoes/types/textblock.h"
 #include "shoes/types/timerbase.h"
 #include "shoes/types/video.h"
 
@@ -42,5 +43,6 @@
 	shoes_text_init(); \
 	shoes_text_link_init(); \
 	shoes_text_view_init(); \
+	shoes_textblock_init(); \
 	shoes_timerbase_init(); \
 	shoes_video_init();
